This permitted to simplify a bit the overall API,
remove the need to check offsets > 1.0f, and made possible
to precise the path move event destination node, which is
a first step toward making it fully work.

The detection algorithm has still issues with wall collision
but the fix is still being studied.

Note that I kept on purpose the pixel rounding to keep avoiding
the tiles flipping, which was the main cause of splitting
the position from their offsets, even if I appeared to be not quite
necessary.

+++ src/modes/map/map_zones.h
@@ -117,7 +117,7 @@ public:
 	*** checked as to whether or not it may be found in this zone, the floating point portion of its map coordinates are not taken
 	*** into account.
 	**/
-	bool IsInsideZone(uint16 pos_x, uint16 pos_y) const;
+	bool IsInsideZone(float pos_x, float pos_y) const;
 
 	//! \name Class member accessor methods
 	//@{
@@ -139,7 +139,7 @@ protected:
 	*** \param x A reference where to store the value of the x position
 	*** \param y A reference where to store the value of the x position
 	**/
-	void _RandomPosition(uint16& x, uint16& y);
+	void _RandomPosition(float& x, float& y);
 }; // class MapZone
